Comcast Consumer complaint status
Comcast status of complaint is consistently increasing during the times so analyzing and seprating the number of complaint  and its order in different location with respect to numbers. finally pointing out specific complaint  number registered as telecom disfunction and the exact number in that region. After the complaint complaint  status for more clarity on complaint open and closed.